+++ admin/include/donvi/xuly.php
@@ -1,5 +1,6 @@
 <?php 
 include("../../config.php");
+session_start();
 $form = "";
 $id = "";
 $tendv = "";
@@ -16,19 +17,36 @@ if(isset($_POST["khoi"])){$khoi = $_POST["khoi"];}
 if(isset($_POST["tt"])){$tt = $_POST["tt"];}
 if(isset($_POST["trangthai"])){$trangthai = $_POST["trangthai"];}
 
+// Lấy thông tin phân quyền
+$sqlphanquyen = "select them, sua, xoa from phanquyen where user = '$_SESSION[user_huye_id]' and form = '$form'";
+$tbphanquyen = mysqli_query($con,$sqlphanquyen);
+$rsphanquyen = mysqli_fetch_array($tbphanquyen);
+
 if(isset($_POST["them"])){
-    $sql = "insert into donvi(tendv, tendaydu, khoi, trangthai, tt ) values('$tendv','$tendaydu','$khoi','$trangthai','$tt')";
-    mysqli_query($con,$sql);
-    header("location: ../../index.php?form=".$form);
+    if($rsphanquyen["them"] == 1){
+        $sql = "insert into donvi(tendv, tendaydu, khoi, trangthai, tt ) values('$tendv','$tendaydu','$khoi','$trangthai','$tt')";
+        mysqli_query($con,$sql);
+        header("location: ../../index.php?form=".$form);
+    }else{
+        header("location: ../../index.php?form=".$form."&false=false");
+    }
 }
 if(isset($_POST["sua"])){
-    $sql = "update donvi set tendv = '$tendv', tendaydu = '$tendaydu', khoi = '$khoi', trangthai = '$trangthai', tt = '$tt' where id = '$id'";
-    mysqli_query($con,$sql);
-    header("location: ../../index.php?form=".$form."&act=edit&id=".$id);
+    if($rsphanquyen["sua"]== 1){
+        $sql = "update donvi set tendv = '$tendv', tendaydu = '$tendaydu', khoi = '$khoi', trangthai = '$trangthai', tt = '$tt' where id = '$id'";
+        mysqli_query($con,$sql);
+        header("location: ../../index.php?form=".$form."&act=edit&id=".$id);
+    }else{
+        header("location: ../../index.php?form=".$form."&false=false");
+    }
 }
 if(isset($_POST["xoa"])){
-    $sql = "delete from donvi where id = '$id'";
-    mysqli_query($con,$sql);
-    header("location: ../../index.php?form=".$form);
+    if($rsphanquyen["xoa"]== 1){
+        $sql = "delete from donvi where id = '$id'";
+        mysqli_query($con,$sql);
+        header("location: ../../index.php?form=".$form);
+    }else{
+        header("location: ../../index.php?form=".$form."&false=false");
+    }
 }
 ?>
